Add Hirst painting project to turtle exercises
This commit introduces a new file for a Hirst painting project which uses colorgram to extract colors from an image and Python turtle graphics to paint a Hirst-style dot grid. The README.md was also updated to reflect this additional project within the turtle exercises section. This new project provides a more complex task showcasing the capabilities of turtle graphic functionality.

Signed-off-by: Arnaud Ober <me@arnaudober.com>