Add Etch a Sketch game functionality
Implemented a simple Etch a Sketch game using the turtle package in Python. This allows the user to control a turtle object to move up, down, left, and right, draw and clear the screen. The README.md file has been updated to include this new addition in the list of exercises. This addition provides a simple and interactive way to get acquainted with event-driven programming in Python.

Signed-off-by: Arnaud Ober <me@arnaudober.com>