Add Turtles Race game
A new game, Turtles Race, has been added to the project. The game allows multiple turtles of different colors to race against each other and the user is allowed to bet on the winning turtle. The game was implemented in a new python file, main.py, in the turtles-race directory.

Signed-off-by: Arnaud Ober <me@arnaudober.com>