Add birthday wisher functionality and template letters
Introduced a new feature "birthday wisher" that reads the birthdays.csv file to determine who has a birthday today. Templates for birthday messages are added and will be sent to the birthday person's email. The quotes functionality has been replaced with a TODO list outlining the steps needed for the birthday wisher.